setup.d: Add script to patch bashrc.
~/.bashrc seems to vary wildly between distros. Kind of strange and a
little frustrating. The aim here is to do as little to bashrc as
possible to avoid surprise.

It would be nice if all distros just had a bashrc that sourced
everything in a user-filled directory, like this change does.

This script means that for every new shell, all of the scripts in
~/.bash.d/ will run. They shouldn't be very complex as such. This
much should be obvious though - why do you want a massively complex
environment for your shell?